stage1

+++ DalFacade/DalApi/IEngineer.cs
@@ -3,10 +3,10 @@ namespace DalApi;
 using DO;
 public interface IEngineer
 {
-    int Create(Engineer item); //Creates new entity object in DAL
-    Engineer? Read(int id); //Reads entity object by its ID
+    long Create(Engineer item); //Creates new entity object in DAL
+    Engineer? Read(long id); //Reads entity object by its ID
     List<Engineer?> ReadAll(); //stage 1 only, Reads all entity objects
     void Update(Engineer item); //Updates entity object
-    void Delete(int id); //Deletes an object by is Id
+    void Delete(long id); //Deletes an object by is Id
 
 }
